fix issue #1708
