+++ net/ghttp/ghttp_z_unit_feature_router_strict_test.go
@@ -19,7 +19,7 @@ import (
 	"github.com/gogf/gf/v2/util/guid"
 )
 
-func Test_Router_Handler_Extended_Handler_WithObject(t *testing.T) {
+func Test_Router_Handler_Strict_WithObject(t *testing.T) {
 	type TestReq struct {
 		Age  int
 		Name string
@@ -130,7 +130,7 @@ func (ControllerForHandlerWithObjectAndMeta2) Test4(ctx context.Context, req *Te
 		Name: req.Name,
 	}, nil
 }
-func Test_Router_Handler_Extended_Handler_WithObjectAndMeta(t *testing.T) {
+func Test_Router_Handler_Strict_WithObjectAndMeta(t *testing.T) {
 	s := g.Server(guid.S())
 	s.Use(ghttp.MiddlewareHandlerResponse)
 	s.Group("/", func(group *ghttp.RouterGroup) {
@@ -152,7 +152,7 @@ func Test_Router_Handler_Extended_Handler_WithObjectAndMeta(t *testing.T) {
 	})
 }
 
-func Test_Router_Handler_Extended_Handler_Group_Bind(t *testing.T) {
+func Test_Router_Handler_Strict_Group_Bind(t *testing.T) {
 	s := g.Server(guid.S())
 	s.Use(ghttp.MiddlewareHandlerResponse)
 	s.Group("/api/v1", func(group *ghttp.RouterGroup) {
@@ -188,5 +188,63 @@ func Test_Router_Handler_Extended_Handler_Group_Bind(t *testing.T) {
 		t.Assert(client.GetContent(ctx, "/api/v2/custom-test2?age=18&name=john"), `{"code":0,"message":"","data":{"Id":1,"Name":"john"}}`)
 		t.Assert(client.GetContent(ctx, "/api/v2/custom-test3?age=18&name=john"), `{"code":0,"message":"","data":{"Id":1,"Age":18}}`)
 		t.Assert(client.GetContent(ctx, "/api/v2/custom-test4?age=18&name=john"), `{"code":0,"message":"","data":{"Id":1,"Name":"john"}}`)
+	})
+}
+
+func Test_Issue1708(t *testing.T) {
+	type Test struct {
+		Name string `json:"name"`
+	}
+	type Req struct {
+		Page       int      `json:"page"       dc:"分页码"`
+		Size       int      `json:"size"       dc:"分页数量"`
+		TargetType string   `json:"targetType" v:"required#评论内容类型错误" dc:"评论类型: topic/ask/article/reply"`
+		TargetId   uint     `json:"targetId"   v:"required#评论目标ID错误" dc:"对应内容ID"`
+		Test       [][]Test `json:"test"`
+	}
+	type Res struct {
+		Page       int      `json:"page"       dc:"分页码"`
+		Size       int      `json:"size"       dc:"分页数量"`
+		TargetType string   `json:"targetType" v:"required#评论内容类型错误" dc:"评论类型: topic/ask/article/reply"`
+		TargetId   uint     `json:"targetId"   v:"required#评论目标ID错误" dc:"对应内容ID"`
+		Test       [][]Test `json:"test"`
+	}
+
+	s := g.Server(guid.S())
+	s.Use(ghttp.MiddlewareHandlerResponse)
+	s.BindHandler("/test", func(ctx context.Context, req *Req) (res *Res, err error) {
+		return &Res{
+			Page:       req.Page,
+			Size:       req.Size,
+			TargetType: req.TargetType,
+			TargetId:   req.TargetId,
+			Test:       req.Test,
+		}, nil
+	})
+	s.SetDumpRouterMap(false)
+	s.Start()
+	defer s.Shutdown()
+
+	time.Sleep(100 * time.Millisecond)
+	gtest.C(t, func(t *gtest.T) {
+		client := g.Client()
+		client.SetPrefix(fmt.Sprintf("http://127.0.0.1:%d", s.GetListenedPort()))
+		content := `
+{
+    "targetType":"topic",
+    "targetId":10785,
+    "test":[
+        [
+            {
+                "name":"123"
+            }
+        ]
+    ]
+}
+`
+		t.Assert(
+			client.PostContent(ctx, "/test", content),
+			`{"code":0,"message":"","data":{"page":0,"size":0,"targetType":"topic","targetId":10785,"test":[[{"name":"123"}]]}}`,
+		)
 	})
 }
